Moved game-related logic to game.cs, fixed reset, refined pause/resume

+++ RogersErwin_Assign5/Stage.cs
@@ -39,7 +39,7 @@ namespace RogersErwin_Assign5
         public List<int> correctColumnSums { get; set; }    // Same, but for above going from left to right.
         public int correctDiagonalSum { get; set; }         // Same, but only for the one diagonal sum.
 
-        public double millisecondsElapsed { get; set; }
+        public long millisecondsElapsed { get; set; }
 
         public bool completed { get; set; }                 // True if this board is solved, or false if in progess.
 
@@ -52,7 +52,7 @@ namespace RogersErwin_Assign5
         }
 
         public Stage(List<int> boardValues, List<int> solutionValues, List<Point> lockedValues, int gameSize, string stageName,
-            List<int> correctRowSums, List<int> correctColumnSums, int correctDiagonalSum, double millisecondsElapsed)
+            List<int> correctRowSums, List<int> correctColumnSums, int correctDiagonalSum, long millisecondsElapsed)
         {
             this.boardValues = boardValues;
             this.solutionValues = solutionValues;
